+++ src/main/java/com/m/blog/domain/posting/infrastructure/repository/PostingDslRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface PostingDslRepository {
-    int findMaxId(int boardCollectionId, int boardId);
+    long findMaxId(long boardCollectionId, long boardId);
     Page<PostingDto> getNonFilteredPage(Pageable pageable);
 
-    Page<PostingDto> getFilteredPage(int boardCollectionIdm, int boardId, Pageable pageable);
+    Page<PostingDto> getFilteredPage(long boardCollectionIdm, long boardId, Pageable pageable);
     PostingDto getSinglePage(Posting.PostingId condition);
 }
